# Speed Up Worst-Case Overhead