--- a/MANIFEST.in
+++ b/MANIFEST.in
@@ -9,6 +9,7 @@ recursive-include requirements *.txt
 recursive-include tests *.py
 recursive-include line_profiler *.txt
 recursive-include line_profiler *.pyx
+recursive-include line_profiler *.pyi
 recursive-include line_profiler *.pxd
 recursive-include line_profiler *.pyd
 recursive-include line_profiler *.c