The following PR transforms this callback to this. More yellow is worse.
Before:
<img width="680" height="789" alt="image" src="https://github.com/user-attachments/assets/718301dd-b8f9-4a8e-b592-c26bd84471b6" />
After:
<img width="788" height="962" alt="image" src="https://github.com/user-attachments/assets/34489e4d-2d13-4dbb-8887-3f5dd1cde835" />

The final result is to lower the worst-case overhead from ~60x to ~40x, which gets closer to the ~25-30x measured in 4.1.3. The average overhead remains at ~2x. Worst-case overhead output below:
```
kernprof -lvz test.py
Time Elapsed without profiling: 90.72ms
Time Elapsed profiling: 3686.82ms
Slowdown from profiling: 3596.1000000000004ms, 40.64x
Wrote profile results to 'test.py.lprof'
Timer unit: 1e-06 s

Total time: 2.78653 s
File: test.py
Function: main at line 3

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
     3                                           def main(profiled=False):
     4         1          2.9      2.9      0.0      start = time.perf_counter()
     5   2000001    1401860.4      0.7     50.3      for x in range(2000000): #2000000
     6   2000000    1384564.3      0.7     49.7          y = x
     7         1         31.1     31.1      0.0      elapsed_ms = round((time.perf_counter()-start)*1000, 2)
     8         1          1.2      1.2      0.0      if profiled:
     9         1         60.8     60.8      0.0          print(f"Time Elapsed profiling: {elapsed_ms}ms")
    10                                               else:
    11                                                   print(f"Time Elapsed without profiling: {elapsed_ms}ms")
    12         1          4.9      4.9      0.0      return elapsed_ms
```