@Theelx 2025-08-06
Huh, it's interesting that the tests fail. On my end, the only error I get when running run_tests.py is in the attached screenshot. Any ideas how I can reproduce the errors in the CI on my local machine so I can fix them?
<img width="1845" height="220" alt="image" src="https://github.com/user-attachments/assets/bedb7fec-aa47-4ad6-be0d-ec35fa0c1431" />