@TTsangSC 2025-08-07
My tests so far:
- `_Py_HashBytes()` requires jumping more hoops to get in 3.13+, and is promoted to a public function `Py_HashBuffer()` in 3.14. This is also why `Build sdist` fails. Here is what I added to `Python_wrapper.h`:
  ```python
  #if PY_VERSION_HEX < 0x030e00a1  // 3.14.0a1
  #   define Py_HashBuffer(b, s) _Py_HashBytes(b, s)
  #   if PY_VERSION_HEX < 0x030d00a1  // 3.13.0a1
  #       ifndef Py_BUILD_CORE
  #           define Py_BUILD_CORE 1
  #       endif
  #       include "internal/pycore_pyhash.h"  // _Py_HashBytes()
  #   else
  #       include "pyhash.h"
  #   endif 
  #endif
  ```
- With this the package does build, but `line_profiler` remains un-importable:
  ```python
  Traceback (most recent call last):
    File [...]"/line_profiler.py", line 20, in <module>
      from ._line_profiler import LineProfiler as CLineProfiler
  ImportError: dlopen([...]/_line_profiler.cpython-313-darwin.so, 0x0002): symbol not found in flat namespace '__Z13_Py_HashBytesPKvl' 
  ```