@Theelx 2025-08-07
Ok, when I try building on 3.13.3 on Linux, I need the following code to get it to build:
```c
#if PY_VERSION_HEX < 0x030e00a1  // 3.14.0a1
#   define Py_HashBuffer(b, s) _Py_HashBytes(b, s)
#   ifndef Py_BUILD_CORE
#       define Py_BUILD_CORE 1
#   endif
#   include "internal/pycore_pyhash.h"  // _Py_HashBytes()
#   undef Py_BUILD_CORE
#endif
```
I'm working on getting it to import and run tests now.