@Theelx 2025-08-08
If you remove the following, does it still error?
```py
        else:
            # this should never matter because time isn't used when this is hit
            # but gcc doesn't know that, so we include it to avoid a compiler warning
            time = 0
```