@TTsangSC 2025-08-08
- Commenting out the retrieval of `block_hash` from `_bh_cache` makes all tests pass. But I just can't for the life of me figure out why that part is problematic... 99% of the cases we're just dealing with a bytestring and it shouldn't be all that complicated. And since code objects are immutable I don't see why using their identities as `unordered_map` keys would cause issues – after all I don't think we reuse object addresses in Python.
- About the `time = 0` thing:
  - Good news: commenting it out does make the garbage timings disappear.
  - Bad news: the tests still passed *with* the garbage timings, so maybe some of those need to be rewritten to be more strict.