@Theelx 2025-08-09
Huh, that's odd. Well in that case I'll just remove the block hash cache and time = 0, and I can try to do more tests to figure out why it's a problem later.