@Theelx 2025-08-09
Ok, well it seems the tests pass fine on Ubuntu and Mac but not Windows... does anyone have a Windows computer on which they can debug why this is occurring?