@Theelx 2025-08-09
Awesome, thanks for the test script and measurements! I currently have a few pending commits locally to extend the iPython/Jupyter notebook support with an lprun_all magic function (to automatically profile the entire cell being run instead of just a specific function/module), but it's based off this branch so the sooner these test issues can be resolved and this PR merged, the better. Once this is merged (or closed), I can submit the new PR (which is intended for students to use this Fall semester). Once that is done, I think I'm going to work on integrating a benchmarking harness and other representative use cases to run on new commits/PRs here so we can measure and detect regressions before they happen.