@TTsangSC 2025-08-10
@Theelx

Just figured that I should maybe also ping you here – apparently people don't necessarily get notifications when their forks get PRs, been there.

In Theelx/line_profiler#4 I replaced those repeated item accesses in `inner_trace_callback()` with pointers and squeezed quite a bit of performance out of them. However, the bad news is that `Py_HashBuffer()` doesn't seem to be any faster than `PyObject_Hash()` (at least in my tests) so those yellow lines may be here to stay.