@Theelx 2025-08-10
Thanks for pinging me, I indeed had not gotten any notifications. When I run the test script you included on the latest version of the code on 3.13.3 (saved as test5.py) and compare it to the original test script I had (saved as test.py), I get very different numbers for slowdowns. Am I calling your script wrong? The latest version of the code (with PyObject_Hash instead of hash_bytecode) is being used in this case, I'm planning to check hash_bytecode next. I'm not sure if this is related, but I think the two different invocations (python3 vs kernprof) might have something to do with this discrepancy (especially because of the different output units). I'm going to try changing your script to run with the cmdline kernprof.

```sh
kernprof -lvz test.py
Time Elapsed without profiling: 91.4ms
Time Elapsed profiling: 2968.16ms
Slowdown from profiling: 2876.7599999999998ms, 32.47x
Wrote profile results to 'test.py.lprof'
Timer unit: 1e-06 s

Total time: 2.60081 s
File: test.py
Function: main at line 3

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
     3                                           def main(profiled=False):
     4         1          2.6      2.6      0.0      start = time.perf_counter()
     5   2000001    1310737.8      0.7     50.4      for x in range(2000000): #2000000
     6   2000000    1289972.1      0.6     49.6          y = x
     7         1         25.2     25.2      0.0      elapsed_ms = round((time.perf_counter()-start)*1000, 2)
     8         1          1.4      1.4      0.0      if profiled:
     9         1         68.9     68.9      0.0          print(f"Time Elapsed profiling: {elapsed_ms}ms")
    10                                               else:
    11                                                   print(f"Time Elapsed without profiling: {elapsed_ms}ms")
    12         1          3.9      3.9      0.0      return elapsed_ms

theel@theel-hp:~/Coding/line_profiler/line_profiler$ python3 test5.py
From 5 run(s) @ 2000000 loops: Overhead = 43.82 ± 0.20x (87.40 ms -> 3916.86 ms)
theel@theel-hp:~/Coding/line_profiler/line_profiler$ python3 test5.py -v
Time elapsed without profiling (#1): 85.48 ms
Time elapsed with profiling (#1): 3908.07 ms
Time elapsed without profiling (#2): 85.74 ms
Time elapsed with profiling (#2): 3958.72 ms
Time elapsed without profiling (#3): 85.97 ms
Time elapsed with profiling (#3): 3902.86 ms
Time elapsed without profiling (#4): 85.95 ms
Time elapsed with profiling (#4): 3949.99 ms
Time elapsed without profiling (#5): 85.74 ms
Time elapsed with profiling (#5): 3977.27 ms
Timer unit: 1e-09 s

Total time: 15.4224 s
File: /home/theel/Coding/line_profiler/line_profiler/test5.py
Function: inner_func at line 15

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
    15                                           def inner_func(n: int) -> float:
    16         5       9857.0   1971.4      0.0      start = perf_counter()
    17  10000005 7856910303.0    785.7     50.9      for x in range(n):
    18  10000000 7565454885.0    756.5     49.1          y = x
    19         5      69276.0  13855.2      0.0      return (perf_counter() - start) * 1000

From 5 run(s) @ 2000000 loops: Overhead = 44.93 ± 0.17x (85.78 ms -> 3939.38 ms)
```